requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
geopandas>=0.13.0
keplergl>=0.3.2
folium>=0.14.0
streamlit-folium>=0.13.0
pillow>=9.0.0
shapely>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
matplotlib>=3.7.0

